Sunshine (jan-2014)
SUNSHINE is a rails application designed with the sheer intention of
helping cancer patients find a suitable blood donor and arrange for a
blood transfusion. It was built using ruby on rails. The application
uses Device gem to manage the user logins and also uses an emailer gems
to send out emails to contact the donors.